Refine DPOR logic minus backtracking

+++ docs/design/include/mcmini/model/defines.hpp
@@ -1,4 +1,4 @@
 #pragma once
 #include <cstdint>
 
-using runner_id_t = uint32_t;
+using runner_id_t = uint16_t;
